+++ package.json
@@ -60,6 +60,7 @@
     "typescript": "^4.7.4"
   },
   "dependencies": {
+    "crypto": "npm:crypto-browserify@^3.12.0",
     "@sentry/react": "^7.48.0",
     "classnames": "^2.3.1"
   }
